remove leaderboard support
as a mostly-singleplayer type-player, i really don't see a valid reason for creating / supporting a leaderboard. as this app also doesn't have web support as i don't want to code / host that, the only leaderboard scores would be you racing yourself to the top, which i think is practically useless.